add installation logic to make default revision if there isn't one
