+++ istioctl/pkg/tag/generate.go
@@ -6,6 +6,7 @@ import (
 	_ "embed"
 	"encoding/base64"
 	"fmt"
+	"io/ioutil"
 	"strings"
 	"text/template"
 
@@ -119,6 +120,14 @@ func Generate(ctx context.Context, client kube.ExtendedClient, opts *GenerateOpt
 	}
 
 	return tagWhYAML, nil
+}
+
+// Create applies the given tag manifests.
+func Create(client kube.ExtendedClient, manifests string) error {
+	if err := applyYAML(client, manifests, "istio-system"); err != nil {
+		return fmt.Errorf("failed to apply tag manifests to cluster: %v", err)
+	}
+	return nil
 }
 
 // generateValidatingWebhook renders a validating webhook configuration from the given tagWebhookConfig.
@@ -275,4 +284,32 @@ func tagWebhookConfigFromValidatingWebhook(wh admit_v1.ValidatingWebhookConfigur
 		CABundle:       caBundle,
 		IstioNamespace: "istio-system",
 	}, nil
+}
+
+// applyYAML taken from remote_secret.go
+func applyYAML(client kube.ExtendedClient, yamlContent, ns string) error {
+	yamlFile, err := writeToTempFile(yamlContent)
+	if err != nil {
+		return fmt.Errorf("failed creating manifest file: %w", err)
+	}
+
+	// Apply the YAML to the cluster.
+	if err := client.ApplyYAMLFiles(ns, yamlFile); err != nil {
+		return fmt.Errorf("failed applying manifest %s: %v", yamlFile, err)
+	}
+	return nil
+}
+
+// writeToTempFile taken from remote_secret.go
+func writeToTempFile(content string) (string, error) {
+	outFile, err := ioutil.TempFile("", "revision-tag-manifest-*")
+	if err != nil {
+		return "", fmt.Errorf("failed creating temp file for manifest: %w", err)
+	}
+	defer func() { _ = outFile.Close() }()
+
+	if _, err := outFile.Write([]byte(content)); err != nil {
+		return "", fmt.Errorf("failed writing manifest file: %w", err)
+	}
+	return outFile.Name(), nil
 }
