+++ istioctl/pkg/tag/templates/validatingwebhook.yaml
@@ -1,7 +1,7 @@
 apiVersion: admissionregistration.k8s.io/v1
 kind: ValidatingWebhookConfiguration
 metadata:
-  name: istiod-default-validator-{{ .IstioNamespace }}
+  name: istiod-default-validator
   labels:
     app: istiod
     istio: istiod
