+++ manifests/charts/istiod-remote/templates/validatingwebhookconfiguration.yaml
@@ -51,38 +51,6 @@ webhooks:
           {{- else }}
           - "{{ .Values.revision }}"
           {{- end }}
-  # Webhook handling default validation
-  - name: validation.istio.io
-    clientConfig:
-      # Should change from base but cannot for API compat
-      {{- if .Values.base.validationURL }}
-      url: {{ .Values.base.validationURL }}
-      {{- else }}
-      service:
-        name: istiod{{- if not (eq .Values.revision "") }}-{{ .Values.revision }}{{- end }}
-        namespace: {{ .Values.global.istioNamespace }}
-        path: "/validate"
-      {{- end }}
-      caBundle: ""
-    rules:
-      - operations:
-          - CREATE
-          - UPDATE
-        apiGroups:
-          - security.istio.io
-          - networking.istio.io
-          - telemetry.istio.io
-        apiVersions:
-          - "*"
-        resources:
-          - "*"
-    failurePolicy: Ignore
-    sideEffects: None
-    admissionReviewVersions: ["v1beta1", "v1"]
-    objectSelector:
-      matchExpressions:
-        - key: istio.io/rev
-          operator: DoesNotExist
 ---
 {{- end }}
 {{- end }}
